requirements.txt
# Core
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
python-dotenv>=1.0.0

# AI
anthropic>=0.42.0

# Job Discovery
google-search-results>=2.4.2  # SerpAPI

# Resume Generation
python-docx>=1.1.0

# Database
aiosqlite>=0.20.0

# Browser Automation (Phase 4)
# playwright>=1.49.0

# Utilities
httpx>=0.28.0
